+++ rpx_pro/api.py
@@ -1,14 +1,12 @@
 """RPXProAPI: Programmatische Python-API fuer RPX Pro."""
 
 import random
-import time
-from typing import Dict, List, Optional, Any
+from typing import List, Optional
 
 from rpx_pro.constants import generate_short_id
 from rpx_pro.models.enums import MessageRole, MissionStatus
 from rpx_pro.models.entities import Character
-from rpx_pro.models.world import World, WorldSettings
-from rpx_pro.models.session import ChatMessage, Mission, Session
+from rpx_pro.models.session import ChatMessage, Mission
 from rpx_pro.managers.data_manager import DataManager
 from rpx_pro.managers.prompt_generator import PromptGenerator